tách api product -> hook
